+++ kite/src/main/kotlin/com/tang/kite/sql/provider/SqlProvider.kt
@@ -63,4 +63,6 @@ interface SqlProvider {
 
     fun <T> paginate(clazz: Class<T>, entity: Any?, orderBys: Array<OrderItem<T>>, pageNumber: Long, pageSize: Long): SqlStatement
 
+    fun <T> paginateWithJoins(clazz: Class<T>, entity: Any?, orderBys: Array<OrderItem<T>>, pageNumber: Long, pageSize: Long, withAlias: Boolean = true): SqlStatement
+
 }